% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedsub.mas.R
\name{seedsub.mas}
\alias{seedsub.mas}
\title{Some exploration of the seedling preferences for various substrates on
Mt. Moosilauke.  This package will just contain data files to be used
in interactive applications.}
\description{
Some exploration of the seedling preferences for various substrates on
Mt. Moosilauke.  This package will just contain data files to be used
in interactive applications.
}

