% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trsubmas11c.R
\docType{data}
\name{trsubmas11c}
\alias{trsubmas11c}
\title{Transect substrate data}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 146 rows and 35 columns.
\itemize{
    \item TRAN: "character" with 9 unique values and 0 NA values.
    \tabular{rrrrrrrrr}{
    \strong{E320} \tab \strong{E330} \tab \strong{E334} \tab \strong{E335} \tab \strong{S344} \tab \strong{W32} \tab \strong{W46} \tab \strong{W60} \tab \strong{W70}\cr
    20 \tab 16 \tab 17 \tab 15 \tab 5 \tab 12 \tab 20 \tab 19 \tab 22
    }
    \item TPLOT: "double" variable with 10 unique values in the range [15, 24], 0 NA values.
    \item QPOS: "character" with 7 unique values and 0 NA values.
    \tabular{rrrrrrr}{
    \strong{""} \tab \strong{A} \tab \strong{B} \tab \strong{C} \tab \strong{L} \tab \strong{R} \tab \strong{U}\cr
    43 \tab 10 \tab 10 \tab 10 \tab 25 \tab 23 \tab 25
    }
    \item CENS: "double" variable with 2 unique values in the range [1999, 2011], 0 NA values.
    \item RCK: "double" variable with 5 unique values in the range [0, 70], 0 NA values.
    \item BSOIL: "double" variable with 2 unique values in the range [0, 0.5], 0 NA values.
    \item WATER: "double" variable with 3 unique values in the range [0, 6], 0 NA values.
    \item BLA5: "double" variable with 14 unique values in the range [0, 20], 0 NA values.
    \item BLD5: "double" variable with 8 unique values in the range [0, 6], 0 NA values.
    \item TIPA: "double" variable with 4 unique values in the range [0, 15], 0 NA values.
    \item STPA: "double" variable with 9 unique values in the range [0, 20], 0 NA values.
    \item LITT: "double" variable with 79 unique values in the range [1, 94.4], 0 NA values.
    \item LITM: "double" variable with 51 unique values in the range [0, 94], 0 NA values.
    \item LITD: "double" variable with 2 unique values in the range [0, 0.5], 0 NA values.
    \item LITC: "double" variable with 61 unique values in the range [0, 94], 0 NA values.
    \item MSSG: "double" variable with 66 unique values in the range [1, 98], 0 NA values.
    \item WDG5: "double" variable with 15 unique values in the range [0, 17], 73 NA values.
    \item WDA5: "double" variable with 18 unique values in the range [0, 25], 73 NA values.
    \item MSWDA5: "double" variable with 10 unique values in the range [0, 10], 73 NA values.
    \item LTWDA5: "double" variable with 4 unique values in the range [0, 3], 73 NA values.
    \item WDG2: "double" variable with 12 unique values in the range [0, 13], 73 NA values.
    \item MSBLA5: "double" variable with 7 unique values in the range [0, 7], 73 NA values.
    \item MSBLD5: "double" variable with 4 unique values in the range [0, 1], 73 NA values.
    \item MSSTP: "double" variable with 6 unique values in the range [0, 35], 73 NA values.
    \item MSRCK: "double" variable with 7 unique values in the range [0, 25], 73 NA values.
    \item MSTIP: "double" variable with 4 unique values in the range [0, 12], 73 NA values.
    \item MSWDG2: "double" variable with 13 unique values in the range [0, 20], 73 NA values.
    \item WDSTP: "double" variable with 2 unique values in the range [0, 0], 73 NA values.
    \item MSWDA2: "double" variable with 12 unique values in the range [0, 20], 73 NA values.
    \item WDA2: "double" variable with 16 unique values in the range [0, 27], 73 NA values.
    \item FIELD: "character" with 3 unique values and 0 NA values.
    \tabular{rrr}{
    \strong{""} \tab \strong{early} \tab \strong{late}\cr
    43 \tab 30 \tab 73
    }
    \item SUMG: "double" variable with 37 unique values in the range [73, 139], 0 NA values.
    \item SUMA: "double" variable with 28 unique values in the range [91, 139], 0 NA values.
    \item CORRECT: "double" variable with 3 unique values in the range [1, 2], 60 NA values.
    \item SUMGC: "double" variable with 4 unique values in the range [99.5, 100.5], 52 NA values.
}}
\usage{
trsubmas11c
}
\description{
Transect substrate data
}
\keyword{datasets}

